requirements.txt
snowflake-snowpark-python
streamlit
langchain
langchain-community
langchain-snowflake
langgraph
python-dotenv
pandas
